requirements.txt
ultralytics==8.2.14
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
optuna>=3.0.0
PyYAML>=6.0
pathlib
